+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [2.5.0-0] - 2025-09-10
 
-- New itertool: `groupBy`
+- New itertools: `groupBy` and `indexBy`
 - Renamed `groupby` to `igroupby`
 - Deprecated `groupby` (now an alias to `igroupby`)